Finished Project
This Library Management System is a web application designed for efficiently managing and organizing a library's book collection. The backend is developed using C# with ASP.NET Core 6.0, implementing CRUD operations and securely connecting to a MySQL database. The frontend is built with React and styled using Bootstrap, resulting in a dynamic, responsive, and visually appealing user interface.